+++ backend/fortran/src/gmsh_post_process.f
@@ -11,20 +11,20 @@ c
 
       use numbatmod
 
-      integer*8 nval, nel, npt, nnodes, plot_val, E_H_field
-      integer*8 nb_typ_el
-      integer*8 table_nod(nnodes,nel), type_el(nel)
-      integer*8 visite(npt)
+      integer(8) nval, nel, npt, nnodes, plot_val, E_H_field
+      integer(8) nb_typ_el
+      integer(8) table_nod(nnodes,nel), type_el(nel)
+      integer(8) visite(npt)
       double precision x(2,npt)
-      complex*16 sol(3,nnodes+7,nval,nel), n_eff(nb_typ_el)
+      complex(8) sol(3,nnodes+7,nval,nel), n_eff(nb_typ_el)
       integer alloc_stat
-      complex*16, dimension(:,:), allocatable :: sol_avg
+      complex(8), dimension(:,:), allocatable :: sol_avg
 
-      complex*16 val_cmplx(nval)
+      complex(8) val_cmplx(nval)
 
 
       double precision xel(3,nnodes_0), xel_p1(3,3)
-      complex*16 sol_el(3,nnodes_0), sol_max(4)
+      complex(8) sol_el(3,nnodes_0), sol_max(4)
       double precision sol_el_abs2(nnodes_0)
       double precision sol_el_abs2_eE(nnodes_0)
 C      double precision sol_el_abs2_iD(nnodes_0)
@@ -33,15 +33,15 @@ C      double precision ls_im_index(nnodes_0), im_index
 C      double precision ls_abs_index(nnodes_0), abs_index
       double precision v_im, v_re
 
-      integer*8 i, j, i1, iel, namelen, namelen2, typ_e
-      integer*8 q_average, plot_imag, plot_real, plot_abs
-      integer*8 debug, ui
-      complex*16 z_tmp1
+      integer(8) i, j, i1, iel, namelen, namelen2, typ_e
+      integer(8) q_average, plot_imag, plot_real, plot_abs
+      integer(8) debug, ui
+      complex(8) z_tmp1
 
       character(len=FNAME_LENGTH) tchar, gmsh_file_pos, dir_name
       character tval*4, buf*3
       character*1 tE_H
-      integer*8 namelength
+      integer(8) namelength
 c
 
 c
