+++ backend/fortran/src/z_indexx.f
@@ -1,15 +1,15 @@
-c     This subroutine takes an array of complex numbers, computes their squared magnitudes, 
-c     and sorts the indices of these magnitudes in descending order using a hybrid quicksort-insertion 
-c     sort algorithm. 
+c     This subroutine takes an array of complex numbers, computes their squared magnitudes,
+c     and sorts the indices of these magnitudes in descending order using a hybrid quicksort-insertion
+c     sort algorithm.
 c     The sorted indices are stored in the indx array, which can be used to
 c     reorder the original array cor for other purposes where sorted order is required.
 
       SUBROUTINE z_indexx(n,arr,indx)
 
-      INTEGER*8 n,indx(n),M,NSTACK
-      complex*16 arr(n)
+      integer(8) n,indx(n),M,NSTACK
+      complex(8) arr(n)
       PARAMETER (M=7,NSTACK=50)
-      INTEGER*8 i,indxt,ir,itemp,j,jstack,k,l,istack(NSTACK)
+      integer(8) i,indxt,ir,itemp,j,jstack,k,l,istack(NSTACK)
       double precision a
       integer :: allocate_status=0
       double precision, dimension(:), allocatable :: arr_0
